Tilføjet nyt project:Grabxml (laver Best Knight Route) omdøbt parsegen -> parsergen samt rettet alle kald til programmet fra diverse projekter Rettet Olm641,Olm627 og Olm637 til så de compilerer uden fejl Justeret build-parameter i flere solutions

+++ parsergen/ParserDemo/TestParser.cpp
@@ -97,7 +97,7 @@ TestParser::~TestParser() {
 }
 
 void TestParser::deleteNodeList() {
-  for(int i = 0; i < m_nodeList.size(); i++) {
+  for(size_t i = 0; i < m_nodeList.size(); i++) {
     delete m_nodeList[i];
   }
   m_nodeList.clear();
